+++ src/main/java/uk/gov/hmcts/et/common/model/ccd/types/responsehub/ET3CaseDetailsLinksStatusesUpdateRequest.java
@@ -1,20 +1,19 @@
-package uk.gov.hmcts.et.common.model.ccd.types;
+package uk.gov.hmcts.et.common.model.ccd.types.responsehub;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
 import lombok.Data;
 import lombok.extern.jackson.Jacksonized;
-import uk.gov.hmcts.et.common.model.ccd.types.responsehub.ET3HubLinksStatuses;
 
 @Data
 @Builder
 @Jacksonized
-public class ET3HubLinksStatusesUpdateRequest {
+public class ET3CaseDetailsLinksStatusesUpdateRequest {
 
     @JsonProperty("case_id")
     private String caseId;
     @JsonProperty("case_type_id")
     private String caseTypeId;
-    @JsonProperty("et3_hub_links_statuses")
-    private ET3HubLinksStatuses et3HubLinksStatuses;
+    @JsonProperty("et3_case_details_links_statuses")
+    private ET3CaseDetailsLinksStatuses et3CaseDetailsLinksStatuses;
 }
